+++ apps/consent-ui/src/components/views/Dashboard/TempPdfButtonsWrapper.tsx
@@ -1,0 +1,49 @@
+/*
+ * Copyright (c) 2024 The Ontario Institute for Cancer Research. All rights reserved
+ *
+ * This program and the accompanying materials are made available under the terms of
+ * the GNU Affero General Public License v3.0. You should have received a copy of the
+ * GNU Affero General Public License along with this program.
+ *  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
+ * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT
+ * SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS;
+ * OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER
+ * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+import urlJoin from 'url-join';
+
+import { getAppConfig } from 'src/config';
+import { ASSETS_PATH, CONSENT_PDFS_PATH } from 'src/constants';
+import { getTranslation, ValidLanguage } from 'src/i18n';
+
+import TempPdfButtons from './TempPdfButtons';
+
+const TempPdfButtonsWrapper = ({ currentLang }: { currentLang: ValidLanguage }) => {
+	const { CONSENT_UI_URL } = getAppConfig();
+	const { translate } = getTranslation(currentLang);
+
+	const studyConsentPdfUrl = urlJoin(
+		ASSETS_PATH,
+		CONSENT_PDFS_PATH,
+		translate('assetUrls', 'studyConsentPdf'),
+	);
+
+	const mockSignatureImage = urlJoin(CONSENT_UI_URL, '/assets/images/placeholder-signature.png');
+
+	return (
+		<TempPdfButtons
+			currentLang={currentLang}
+			studyConsentPdfUrl={studyConsentPdfUrl}
+			mockSignatureImage={mockSignatureImage}
+		/>
+	);
+};
+
+export default TempPdfButtonsWrapper;
