+++ apps/consent-ui/src/components/views/Dashboard/index.tsx
@@ -19,7 +19,6 @@
 
 import clsx from 'clsx';
 import Image from 'next/image';
-import urlJoin from 'url-join';
 
 import Card from 'src/components/common/Card';
 import { getTranslation, ValidLanguage } from 'src/i18n';
@@ -27,12 +26,11 @@ import ConsentsImage from 'src/../public/assets/images/consents.jpeg';
 import PaddedContainer from 'src/components/common/PaddedContainer';
 import LocalizedLink from 'src/components/common/Link/LocalizedLink';
 import { getAppConfig } from 'src/config';
-import { ASSETS_PATH, CONSENT_PDFS_PATH } from 'src/constants';
 import { getNotificationTranslations } from 'src/components/providers/NotificationProvider/getNotificationTranslations';
 
 import styles from './Dashboard.module.scss';
 import DashboardNotification from './notifications/DashboardNotification';
-import TempPdfButtons from './TempPdfButtons';
+import TempPdfButtonsWrapper from './TempPdfButtonsWrapper';
 
 // TODO remove after backend hookup
 const statuses = ['disabled', 'incomplete', 'complete'] as const;
@@ -41,12 +39,6 @@ const consentStatus = statuses[2];
 const DashboardComponent = async ({ currentLang }: { currentLang: ValidLanguage }) => {
 	const { translate } = getTranslation(currentLang);
 	const notificationTranslations = getNotificationTranslations(currentLang);
-
-	const studyConsentPdfUrl = urlJoin(
-		ASSETS_PATH,
-		CONSENT_PDFS_PATH,
-		translate('assetUrls', 'studyConsentPdf'),
-	);
 
 	const { FEATURE_CONSENT_PDF_BUTTONS } = getAppConfig();
 
@@ -63,9 +55,7 @@ const DashboardComponent = async ({ currentLang }: { currentLang: ValidLanguage 
 				<div className={styles.content}>
 					<h2>{translate('dashboard', 'reviewOhcrnConsents')}</h2>
 					<p>{translate('dashboard', 'reviewConsentsDescription')}</p>
-					{FEATURE_CONSENT_PDF_BUTTONS && (
-						<TempPdfButtons currentLang={currentLang} studyConsentPdfUrl={studyConsentPdfUrl} />
-					)}
+					{FEATURE_CONSENT_PDF_BUTTONS && <TempPdfButtonsWrapper currentLang={currentLang} />}
 					<div className={styles['button-container']}>
 						{consentStatus == 'complete' ? (
 							<LocalizedLink
